elasticsearch-vs-vector-search/requirements.txt
elasticsearch==8.11.0
psycopg2-binary==2.9.9
pgvector==0.2.4
openai==1.3.5
numpy==1.26.2
pandas==2.1.3
sentence-transformers==2.2.2
requests==2.31.0
python-dotenv==1.0.0
rich==13.7.0
tabulate==0.9.0
